// packages/spreadsheet-component/src/types/xlsx.ts
export interface XLSXFileStructure {
  sheets: XLSXImportFile[];
  workbook: XLSXImportFile;
  styles: XLSXImportFile;
  sharedStrings: XLSXImportFile;
  theme?: XLSXImportFile;
  charts: XLSXImportFile[];
  figures: XLSXImportFile[];
  tables: XLSXImportFile[];
  externalLinks: XLSXImportFile[];
  images: XLSXImageFile[];
}
// ...
export interface XLSXTheme {
  clrScheme?: XLSXColorScheme[];
}

export interface XLSXColorScheme {
  name: string;
  value: string;
  lastClr?: string;
}

export interface XLSXNumFormat {
  id: number;
  format: Format;
}
export interface XLSXBorder {
  top?: XLSXBorderDescr;
  left?: XLSXBorderDescr;
  bottom?: XLSXBorderDescr;
  right?: XLSXBorderDescr;
  diagonal?: XLSXBorderDescr;
  diagonalUp?: boolean;
  diagonalDown?: boolean;
}

export interface XLSXBorderDescr {
  style: XLSXBorderStyle;
  color: XLSXColor;
}

export type ExcelIconSet =
  | "NoIcons"
  | "3Arrows"
  | "3ArrowsGray"
  | "3Symbols"
  | "3Symbols2"
  | "3Signs"
  | "3Flags"
  | "3TrafficLights1"
  | "3TrafficLights2"
  | "4Arrows"
  | "4ArrowsGray"
  | "4RedToBlack"
  | "4Rating"
  | "4TrafficLights"
  | "5Arrows"
  | "5ArrowsGray"
  | "5Rating"
  | "5Quarters"
  | "3Stars"
  | "3Triangles"
  | "5Boxes";

/**
 * Standardized XLSX hexadecimal color (with or without alpha channel).
 * Note that the alpha channel goes first! AARRGGBB
 * e.g. "1E5010" or "331E5010"
 */
export type XlsxHexColor = string & Alias;

export interface ImportedFiles {
  [path: string]:
    | string
    | {
        imageSrc: string;
      };
}

export interface XLSXXmlDocuments {
  [path: string]: XMLDocument;
}

export interface XLSXImageFile {
  fileName: string;
  imageSrc: string;
}
// ...
export type XLSXCfType =
  | "aboveAverage"
  | "expression"
  | "cellIs"
  | "colorScale"
  | "dataBar"
  | "iconSet"
  | "top10"
  | "uniqueValues"
  | "duplicateValues"
  | "containsText"
  | "notContainsText"
  | "beginsWith"
  | "endsWith"
  | "containsBlanks"
  | "notContainsBlanks"
  | "containsErrors"
  | "notContainsErrors"
  | "timePeriod";

export type XLSXCfOperatorType =
  | "beginsWith"
  | "between"
  | "containsText"
  | "endsWith"
  | "equal"
  | "greaterThan"
  | "greaterThanOrEqual"
  | "lessThan"
  | "lessThanOrEqual"
  | "notBetween"
  | "notContains"
  | "notEqual";

export type XLSXHorizontalAlignment =
  | "general"
  | "left"
  | "center"
  | "right"
  | "fill"
  | "justify"
  | "centerContinuous"
  | "distributed";

export type XLSXVerticalAlignment = "top" | "center" | "bottom" | "justify" | "distributed";

export type XLSXCfValueObjectType = "num" | "percent" | "max" | "min" | "percentile" | "formula";

export interface XLSXCfValueObject {
  type: XLSXCfValueObjectType;
  gte?: boolean;
  value?: string;
}
export interface XLSXColorScale {
  colors: XLSXColor[];
  cfvos: XLSXCfValueObject[];
}

export interface XLSXIconSet {
  iconSet: ExcelIconSet;
  cfvos: XLSXCfValueObject[];
  cfIcons?: XLSXCfIcon[]; // Icons can be defined individually instead of following iconSet
  showValue?: boolean;
  percent?: boolean;
  reverse?: boolean;
  custom?: boolean;
}

export interface XLSXCfIcon {
  iconSet: ExcelIconSet;
  iconId: number;
}
export interface XLSXConditionalFormat {
  cfRules: XLSXCfRule[];
  sqref: string[];
  pivot?: boolean;
}

export interface XLSXCfRule {
  type: XLSXCfType;
  priority: number;
  formula?: string[];
  colorScale?: XLSXColorScale;
  dataBar?: any;
  iconSet?: XLSXIconSet;
  dxfId?: number;
  stopIfTrue?: boolean;
  aboveAverage?: boolean;
  percent?: boolean;
  bottom?: boolean;
  operator?: XLSXCfOperatorType;
  text?: string;
  timePeriod?: any;
  rank?: number;
  stdDev?: number;
  equalAverage?: boolean;
}

export interface XLSXSharedFormula {
  formula: string;
  refCellXc: string;
}

export interface XLSXColor {
  auto?: boolean;
  indexed?: number;
  rgb?: string;
  tint?: number;
}
// ...
export interface XLSXImportFile {
  file: XMLFile;
  rels?: XMLFile;
}

export interface XMLFile {
  fileName: string;
  xml: XMLDocument;
}